refacto(routing): passage.py devient un paquet
1263 lignes en un seul module, decoupees selon l'ordre dans lequel un
passage se calcule, pour qu'on puisse le lire de haut en bas :

  constants  les nombres de reglage, et pourquoi chacun vaut ce qu'il vaut
  models     SegmentReport, PassageReport, EtaPassagePlan, NoModelCoveredError
  physics    geometrie polaire, derate de houle, regle moteur, courant. Pur
  sampling   decoupe de route, instants d'echantillonnage, fetch et repli
  engine     un troncon, la marche, les avertissements
  single     estimate_passage, estimate_passage_for_arrival
  sweep      estimate_passage_windows, resolve_sweep_interval

Deplacement pur : le corps de chaque fonction est repris a l'octet pres. La
comparaison ligne a ligne entre l'ancien module et le paquet ne fait
apparaitre que des imports, des docstrings de module, __all__, et le seul
ajout de code delibere ci-dessous.

__init__.py re-exporte tout ce que le reste du code importe aujourd'hui, y
compris _layout_speed_kn que les tests de passage lisent, donc aucun import
externe ne change. Verifie par les suites api et mcp-core, inchangees.

Seul ajout de code : resolve_fetch_adapter() dans sampling, unique endroit ou
le moteur construit un OpenMeteoAdapter. Sans lui la construction existait
en deux exemplaires, et le filet des goldens REST, qui remplace ce nom pour
qu'une regression ne compose pas Open-Meteo depuis la CI, n'en couvrait plus
qu'un apres la decoupe. Le filet vise maintenant sampling et un test verifie
qu'il mord vraiment, pour les deux points d'entree.

Goldens inchanges : 34 rapports de caracterisation, 22 goldens REST, 3
goldens MCP, aucune regeneration.
